refactor(weather): drop server action  and implement new behavior
