+++ src/app/(insert-ad-page)/weather/components/client/bottomSection/RecentPlace.tsx
@@ -1,13 +1,13 @@
 "use client"
 import { keyframes, styled } from "@pigment-css/react"
 import { useEffect, useState, useTransition } from "react"
-import { getRecentPlaceData } from "../../../action"
 import { CityCardType, getCityCardList } from "../../../utils/getCityCardList"
 import WeatherAnimation from "./WeatherAnimation"
 import { getWeatherType } from "../../../utils/getWeatherType"
 import { WEATHER_DETAIL } from "../../../constant"
 import { usePathname, useRouter } from "next/navigation"
 import { CityImage } from "../CityImage"
+import { getRecentPlaceData } from "../../../utils/getRecentPlaceData"
 
 const pulse = keyframes({
 	"0%": { backgroundColor: "#eee" },
